Fix: Provide i64 externally

+++ src/builder.rs
@@ -45,7 +45,7 @@ impl<'a> Builder<'a> {
 
     pub fn constant<U>(&self, t: types::Type, v: U) -> Option<Value> {
         Some(Value::new(match t {
-            types::I64 => self.inst_builder.ins().iconst(t, i64::from(v)),
+            types::I64 => self.inst_builder.ins().iconst(t, v),
             types::B1  => self.inst_builder.ins().bconst(t, v),
             _ => return None
         }, t))
